fix: format currency

+++ app/helpers.php
@@ -1,7 +1,7 @@
 <?php
 
 if (! function_exists('formatCurrency')) {
-    function formatCurrency(int $amount): string
+    function formatCurrency(float $amount): string
     {
         return $amount . ' €';
     }
